refactor(halley-wl): centralize pointer, monitor, IPC, and runtime maintenance plumbing
Clamp pointer coordinates to the nearest monitor edge so mixed-height outputs no longer produce
dead-space local coordinates that break panning, hover, or tty cursor placement, and keep
focus-follows-mouse suppressed while a layer-shell surface still owns keyboard focus. Extract
shared monitor fallback, pointer screen-context, and pointer focus helpers, split IPC handling
into node/trail/cluster/monitor modules, and factor runtime maintenance deadline plus backend
hover-target resolution into clearer subsystem-owned helpers to reduce coupling and raise
cohesion without changing behavior.

cargo test -p halley-wl
