+++ crates/halley-wl/src/compositor/focus/system.rs
@@ -1,6 +1,7 @@
 use super::*;
 use crate::compositor::focus::read;
 use crate::compositor::interaction::state::ViewportPanAnim;
+use crate::compositor::surface_ops::stack_focus_target_for_node;
 use eventline::debug;
 use smithay::reexports::wayland_protocols::xdg::shell::server::xdg_toplevel;
 use smithay::reexports::wayland_server::{Resource, protocol::wl_surface::WlSurface};
@@ -11,6 +12,7 @@ use smithay::wayland::selection::primary_selection::set_primary_focus;
 use crate::compositor::ctx::FocusCtx;
 use smithay::input::Seat;
 use std::ops::{Deref, DerefMut};
+use std::time::{Duration, Instant};
 
 pub(crate) fn on_seat_focus_changed(
     ctx: &mut FocusCtx<'_>,
@@ -68,6 +70,18 @@ pub(crate) fn update_selection_focus_from_surface(st: &Halley, surface: Option<&
         client.clone(),
     );
     set_primary_focus(&st.platform.display_handle, &st.platform.seat, client);
+}
+
+pub(crate) fn focus_pointer_target(
+    st: &mut Halley,
+    node_id: NodeId,
+    hold_ms: u64,
+    now: Instant,
+) -> NodeId {
+    let focus_target = stack_focus_target_for_node(st, node_id).unwrap_or(node_id);
+    st.set_recent_top_node(focus_target, now + Duration::from_millis(1200));
+    st.set_interaction_focus(Some(focus_target), hold_ms, now);
+    focus_target
 }
 
 pub(crate) fn surface_is_fully_visible_on_monitor(st: &Halley, monitor: &str, id: NodeId) -> bool {
@@ -147,13 +161,7 @@ impl<T: DerefMut<Target = Halley>> FocusSystemController<T> {
                 .get(&fid)
                 .copied()
             {
-                let target_monitor = self
-                    .model
-                    .monitor_state
-                    .node_monitor
-                    .get(&fid)
-                    .cloned()
-                    .unwrap_or_else(|| self.model.monitor_state.current_monitor.clone());
+                let target_monitor = self.monitor_for_node_or_current(fid);
                 if let Some(space) = self.model.monitor_state.monitors.get_mut(&target_monitor) {
                     space.viewport.center = entry.viewport_center;
                     space.camera_target_center = entry.viewport_center;
