+++ crates/halley-wl/src/input/pointer/mod.rs
@@ -1,5 +1,6 @@
 pub(crate) mod axis;
 pub(crate) mod button;
+mod context;
 pub(crate) mod focus;
 pub(crate) mod motion;
 pub(crate) mod resize;
